refactor: update modals
