+++ src/components/track/TrackActions.tsx
@@ -1,20 +1,11 @@
-import { lazy, useState } from 'react';
 import { Pencil, Upload, Trash } from 'lucide-react';
 
-import { ModalState, ModalStateSchema } from '@/types';
 import { Button } from '../ui/button';
-import { LazyModal } from '../modals/LazyModal';
 import { useTrackContext } from './TrackItemContext';
 
-const ModalTrackUpload = lazy(
-  () => import('@/components/modals/ModalTrackUpload')
-);
-const ModalTrackUpdate = lazy(
-  () => import('@/components/modals/ModalTrackUpdate')
-);
-const ModalTrackDelete = lazy(
-  () => import('@/components/modals/ModalTrackDelete')
-);
+import UpdateTrackModal from '@/components/modals/UpdateTrackModal';
+import UploadTrackModal from '@/components/modals/UploadTrackModal';
+import DeleteTrackModal from '@/components/modals/DeleteTrackModal';
 
 interface TrackActionsProps {
   genres: string[] | undefined;
@@ -23,88 +14,49 @@ interface TrackActionsProps {
 export function TrackActions({ genres }: TrackActionsProps) {
   const { track } = useTrackContext();
 
-  const [openModalDelete, setOpenModalDelete] = useState<ModalState>(
-    ModalStateSchema.Enum.idle
-  );
-  const [openModalUpload, setOpenModalUpload] = useState<ModalState>(
-    ModalStateSchema.Enum.idle
-  );
-  const [openModalUpdate, setOpenModalUpdate] = useState<ModalState>(
-    ModalStateSchema.Enum.idle
-  );
-
   return (
     <div className="flex gap-2">
-      <Button
-        variant="ghost"
-        size="icon"
-        className="text-muted-foreground hover:text-primary focus-within:text-primary"
-        aria-label="Edit track"
-        onClick={() => setOpenModalUpdate(ModalStateSchema.Enum.open)}
-        disabled={track.id.includes('optimistic')}
-        aria-disabled={track.id.includes('optimistic')}
-        data-testid={`edit-track-${track.id}`}
-      >
-        <Pencil className="w-4 h-4" />
-      </Button>
+      <UpdateTrackModal track={track} genres={genres}>
+        <Button
+          variant="ghost"
+          size="icon"
+          className="text-muted-foreground hover:text-primary focus-within:text-primary"
+          aria-label="Edit track"
+          disabled={track.id.includes('optimistic')}
+          aria-disabled={track.id.includes('optimistic')}
+          data-testid={`edit-track-${track.id}`}
+        >
+          <Pencil className="w-4 h-4" />
+        </Button>
+      </UpdateTrackModal>
 
-      <Button
-        variant="ghost"
-        size="icon"
-        className="text-muted-foreground hover:text-primary focus-within:text-primary"
-        aria-label="Upload track"
-        onClick={() => setOpenModalUpload(ModalStateSchema.Enum.open)}
-        disabled={track.id.includes('optimistic')}
-        aria-disabled={track.id.includes('optimistic')}
-        data-testid={`upload-track-${track.id}`}
-      >
-        <Upload className="w-4 h-4" />
-      </Button>
+      <UploadTrackModal track={track}>
+        <Button
+          variant="ghost"
+          size="icon"
+          className="text-muted-foreground hover:text-primary focus-within:text-primary"
+          aria-label="Upload track"
+          disabled={track.id.includes('optimistic')}
+          aria-disabled={track.id.includes('optimistic')}
+          data-testid={`upload-track-${track.id}`}
+        >
+          <Upload className="w-4 h-4" />
+        </Button>
+      </UploadTrackModal>
 
-      <Button
-        variant="ghost"
-        size="icon"
-        className="text-muted-foreground hover:text-primary focus-within:text-primary"
-        aria-label="Delete track"
-        onClick={() => setOpenModalDelete(ModalStateSchema.Enum.open)}
-        disabled={track.id.includes('optimistic')}
-        aria-disabled={track.id.includes('optimistic')}
-        data-testid={`delete-track-${track.id}`}
-      >
-        <Trash className="w-4 h-4" />
-      </Button>
-
-      {openModalUpdate !== ModalStateSchema.Enum.idle && (
-        <LazyModal>
-          <ModalTrackUpdate
-            track={track}
-            genres={genres}
-            open={openModalUpdate}
-            setOpen={setOpenModalUpdate}
-          />
-        </LazyModal>
-      )}
-
-      {openModalUpload !== ModalStateSchema.Enum.idle && (
-        <LazyModal>
-          <ModalTrackUpload
-            track={track}
-            disableAutoFocus={true}
-            open={openModalUpload}
-            setOpen={setOpenModalUpload}
-          />
-        </LazyModal>
-      )}
-
-      {openModalDelete !== ModalStateSchema.Enum.idle && (
-        <LazyModal>
-          <ModalTrackDelete
-            track={track}
-            open={openModalDelete}
-            setOpen={setOpenModalDelete}
-          />
-        </LazyModal>
-      )}
+      <DeleteTrackModal track={track}>
+        <Button
+          variant="ghost"
+          size="icon"
+          className="text-muted-foreground hover:text-primary focus-within:text-primary"
+          aria-label="Delete track"
+          disabled={track.id.includes('optimistic')}
+          aria-disabled={track.id.includes('optimistic')}
+          data-testid={`delete-track-${track.id}`}
+        >
+          <Trash className="w-4 h-4" />
+        </Button>
+      </DeleteTrackModal>
     </div>
   );
 }
